requirements.txt
torch==1.0.0
argcomplete==1.9.4
attrs==18.1.0
numpy==1.15.4
pytest==4.0.2
wanb==0.6.33
